requirements.txt
pillow
pydicom
numpy